Rename schematic worker
